### 2026-03-28 #762 @emilyriederer
Hi @jeremylfarrell ! Thank you for your submission. It's great to see all the checks pass on the first time.

I'm the rotating editor in chief at the moment. You'll have to be patient with me as this is not my field, but I'd love to go a bit deeper on the "best in class" question. I appreciate your detailed comparison of related packages. It's clear there are multiple packages with some overlap but nothing that handles your package end-to-end.

Do you have any commentary on the pros/cons/feasibility of extending the functionality of existing packages versus introducing another entrant? Neither is the "right" answer, but we always like to probe a bit here to foster connection across scientific software.
