### 2026-03-28 #762 @jeremylfarrell
Great questions! I didn't really consider extending existing packages before starting lakefetch, but now that i have built it I have think it belongs as a standalone package, the others are good but not quite the same:

**fetchR/windfetch** (same maintainer, Blake Seers) are architecturally built around coastal/marine workflows, they expect user-supplied coastline polygons and produce fetch for a single water body. The main point of lakefetch is an complete freshwater workflow.

**lakemorpho** (Jeff Hollister, EPA and editor at rOpenSci!) calculates fetch as one metric among many morphometry parameters (shoreline development, max length/width, volume). Its fetch implementation is a single-point, single-lake calculation, great for morphometric characterization but not designed for the multi-site, multi-lake ecological sampling workflow that lakefetch targets.

**waver** uses a fundamentally different algorithm (wedge-based rather than ray-casting) designed for coastal wave exposure indices. The algorithmic difference reflects different use cases.

Overall, the overlap between these packages is similar in concept "measuring distance across water," but the target users and scientific applications are distinct enough that extending any one of them would have required rewriting most of the target packages while potentially disrupting its existing user base.

That said, I'd welcome opportunities for neighborly interoperability, maybe export in formats compatible with lakemorpho workflows or something along those lines. Happy to explore that during review.
