### 2026-04-15 #762 @Pakillo
Hi @jeremylfarrell !

Looking forward to working with you on this package.

I've been doing the initial editor checks and I already have a couple of suggestions before starting peer review:

1. Please make a website documenting your package (see the [devguide](https://devguide.ropensci.org/pkg_building.html#website)). That will greatly help reviewers (and users) to better understand your package. Creating the website should be pretty straightforward with {pkgdown}, but please let me know if you needed help with that.

2. Some examples take a long time to run. I see they are wrapped in `\donttest` but CRAN still runs those sometimes, and I see that is causing sporadic errors due to excessive checktime. I'd recommend simplifying those examples if possible (e.g. smaller dataset), or using `\dontrun` instead, or the roxygen2 `@examplesIf interactive()` approach (only for long examples). See [here](https://contributor.r-project.org/cran-cookbook/general_issues.html#structuring-of-examples) and [here](https://blog.thecoatlessprofessor.com/programming/r/api-packages-and-cran-requirements/).

Please let me know when you are done with that and we can move on with the review.

Cheers
