### 2026-04-16 fc59b77
Address rOpenSci editor feedback: pkgdown site + fix examples
1. Replace all \donttest{} with @examplesIf interactive() to prevent
   CRAN from running long-running examples that require internet
2. Set up pkgdown website with _pkgdown.yml config and GitHub Actions
   workflow for automated deployment to GitHub Pages
3. Add pkgdown site URL to DESCRIPTION and badge to README

Co-Authored-By: Claude Opus 4.6 <noreply@anthropic.com>
