+++ man/assign_sites_to_lakes.Rd
@@ -20,7 +20,7 @@ sf object with sites and added columns for lake_osm_id, lake_name, lake_area_km2
 Perform spatial join to assign each site to its containing lake polygon.
 }
 \examples{
-\donttest{
+\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
 csv_path <- system.file("extdata", "sample_sites.csv", package = "lakefetch")
 sites <- load_sites(csv_path)
 lake_data <- get_lake_boundary(sites)
@@ -34,6 +34,5 @@ sites_assigned <- assign_sites_to_lakes(
 
 # Check assignments
 table(sites_assigned$lake_name)
-}
-
+\dontshow{\}) # examplesIf}
 }
