### 2026-04-16 51185ed
Add ropensci-software-review branch to pkgdown workflow triggers
Co-Authored-By: Claude Opus 4.6 <noreply@anthropic.com>
