### 2026-04-16 f3d60a2
Add lakefetch-package topic to pkgdown reference index
Co-Authored-By: Claude Opus 4.6 <noreply@anthropic.com>
