### 2026-04-16 d2b733e
Deploying to gh-pages from @ jeremylfarrell/lakefetch@f3d60a205a89f18a4a834e6584de02d7ef5ca725 🚀
