+++ articles/index.md
@@ -1,0 +1,8 @@
+# Articles
+
+### All vignettes
+
+- [Getting Started with
+  lakefetch](https://jeremylfarrell.github.io/lakefetch/articles/getting-started.md):
+- [Validation of lakefetch
+  Calculations](https://jeremylfarrell.github.io/lakefetch/articles/validation.md):
