+++ articles/validation.md
@@ -1,0 +1,174 @@
+# Validation of lakefetch Calculations
+
+## Overview
+
+The `lakefetch` package has been validated through three complementary
+approaches:
+
+1.  **Analytical validation** - Testing against synthetic lakes with
+    known geometry
+2.  **Literature validation** - Comparing results to lakes with
+    published morphometry
+3.  **Edge case validation** - Testing challenging geometries (islands,
+    complex shorelines)
+
+All validation scripts are included in the package under
+`inst/validation/`.
+
+## Analytical Validation
+
+Analytical validation uses synthetic lake geometries where the expected
+fetch values can be calculated mathematically. This provides the most
+rigorous test of algorithm correctness.
+
+### Test Cases
+
+| Test                 | Geometry                        | Expected              | Result          |
+|----------------------|---------------------------------|-----------------------|-----------------|
+| Circular lake center | r=1000m, point at center        | 1000m all directions  | PASS (0% error) |
+| Circular lake edge   | r=1000m, point 800m from center | Min: 200m, Max: 1800m | PASS (0% error) |
+| Rectangular lake N/S | 2000x1000m, center point        | 500m                  | PASS (0% error) |
+| Rectangular lake E/W | 2000x1000m, center point        | 1000m                 | PASS (0% error) |
+| Effective fetch      | Circular center, top 3 mean     | 1000m                 | PASS (0% error) |
+
+### Running Analytical Validation
+
+``` r
+source(system.file("validation", "validate_fetch.R", package = "lakefetch"))
+```
+
+## Literature Validation
+
+Literature validation compares fetch calculations against real lakes
+with published morphometry data from the World Lake Database (ILEC) and
+state limnology surveys.
+
+### Test Lakes
+
+| Lake         | State | Published Area | Calculated Area | Result |
+|--------------|-------|----------------|-----------------|--------|
+| Lake Sunapee | NH    | 16.7 km²       | 16.85 km²       | PASS   |
+| Cayuga Lake  | NY    | 172 km²        | 173.87 km²      | PASS   |
+| Green Lake   | WI    | 29.6 km²       | 32.66 km²       | PASS   |
+
+### Validation Approach
+
+For each lake:
+
+1.  Download boundary from OpenStreetMap
+2.  Calculate fetch at the lake centroid
+3.  Compare max fetch to half the published maximum length
+4.  Compare mean fetch to expected range based on lake shape
+
+Cayuga Lake (a Finger Lake) correctly hit the 50 km fetch cap,
+demonstrating proper handling of very elongated lakes.
+
+### Running Literature Validation
+
+``` r
+source(system.file("validation", "validate_literature.R", package = "lakefetch"))
+```
+
+## Edge Case Validation
+
+Edge case validation tests challenging geometries that could cause
+algorithm failures.
+
+### Test Cases
+
+| Test              | Geometry                         | Key Finding                    | Result  |
+|-------------------|----------------------------------|--------------------------------|---------|
+| Central island    | Island at lake center            | Rays correctly stop at island  | PASS    |
+| Offset island     | Island 800m north of center      | North rays blocked at 500m     | PASS    |
+| Complex shoreline | Sinusoidal variation ±300m       | Fetch varies 1119-1881m        | PASS    |
+| Multiple islands  | 3 islands at different positions | Each blocks rays independently | PASS    |
+| Very large lake   | Lake Erie                        | OSM limitation\*               | SKIPPED |
+
+\*Great Lakes require pre-downloaded shapefiles due to OSM bounding box
+limitations.
+
+### Key Validations
+
+- **Islands work correctly**: Fetch rays stop at island boundaries, not
+  pass through
[…]

[…]
