### 2026-04-16 #762 @jeremylfarrell
Hi Francisco,
Thanks for the suggestions!
The pkgdown site is live at https://jeremylfarrell.github.io/lakefetch/.

I've replaced the long run examples in \donttest{} blocks with @examplesIf interactive() throughout the package. This applies to all functions that require an internet connection (OSM downloads, NHD, weather API), so they will only run in interactive sessions and not during automated checks.
Thanks
Jeremy
