### 2026-04-21 #762 @Pakillo
Thank you, Jeremy. The website looks great, as is the documentation. Everything looks very well explained.

I have tried the [example workflow](https://jeremylfarrell.github.io/lakefetch/index.html#example-workflow) in the start page (README) and am getting an error in `plot_fetch_rose` (it looks like site must be the site name, not a number?), and some warnings that some sites could not be matched to a lake. The latter might be intentional? But I think it may be confusing for newcomers in an example, introductory workflow. I would ensure the example workflow runs smoothly and then show elsewhere other difficulties like sites falling off lakes. Could you please fix these two things before moving forward?

Reproducible example:

``` r
library(lakefetch)
#> lakefetch 0.1.3 loaded
#>   Install 'nhdplusTools' for outlet/inlet detection

# Built-in example data
data("adirondack_sites")

# Download lake boundaries from OpenStreetMap
lake <- get_lake_boundary(adirondack_sites)
#> Converting to spatial format...
#> Spherical geometry (s2) switched off
#> Downloading lake boundaries from OpenStreetMap...
#>   Bounding box: [-74.7115, 43.7742, -74.3775, 44.2712]
#>     Trying name-filtered query for: Blue Mountain Lake, Raquette Lake, Long Lake, Tupper Lake
#>     Found 29 polygons
#>     Found 1 multipolygons
#>     Found 23 polygons
#>     Found 1 multipolygons
#>     Found 17 polygons
#>     Found 1 multipolygons
#>     Found 35 polygons
#>     Found 2 multipolygons
#>     Name-filtered query found results, skipping broad query
#>   Combining results...
#>   Total water bodies found: 109
#>   Auto-detected UTM Zone: 18N
#>   Transforming to UTM for analysis...
#>   Processing lake polygons...
#>   Filtered 9 water bodies < 1e-04 km2
#>   Processed 100 unique lake polygons
#>   Total area: 84.69 km2
#> Spherical geometry (s2) switched on

# Calculate fetch with depth estimation
results <- fetch_calculate(sites = adirondack_sites, lake = lake)
#> Effective fetch method: top3
#> Using default depth: 10 m
#> Assigning sites to lakes...
#>   Checking direct intersections...
#>     2 sites matched directly
#>   Checking 10 unmatched sites with 100m tolerance...
#>     1 additional sites matched within tolerance
#>   Trying name-based matching using 'lake.name' column...
#>     Matched 2 sites to 'Blue Mountain Lake' via name matching from 'Blue Mountain Lake'
#>     Matched 2 sites to 'Raquette Lake' via name matching from 'Raquette Lake'
#> Warning in assign_sites_to_lakes(sites_sf, all_lakes, tolerance_m =
#> get_opt("gps_tolerance_m")): Skipped 1 site(s) named 'Raquette Lake': too far
#> from OSM lake 'Raquette Lake' (>500m). Site(s) may be on a different water
#> body. Sites: Raquette_3
#>     Matched 1 sites to 'Long Lake' via name matching from 'Long Lake'
#>     Matched 1 sites to 'Tupper Lake' via name matching from 'Tupper Lake'
#> Warning in assign_sites_to_lakes(sites_sf, all_lakes, tolerance_m =
#> get_opt("gps_tolerance_m")): Skipped 2 site(s) named 'Tupper Lake': too far
#> from OSM lake 'Tupper Lake' (>500m). Site(s) may be on a different water body.
#> Sites: Tupper_2, Tupper_3
#>   Site assignment summary:
#>     Matched: 9/12
#>     Unmatched: 3
#>     Unmatched sites claim to be in: Raquette Lake, Tupper Lake
#>     Unmatched coordinate range:
#>       Lat: 4852677.9675 to 4896951.5526
#>       Lon: 526526.3106 to 543949.4583
#>     TIP: Try lakefetch_options(gps_tolerance_m = 100) for larger buffer
#>     TIP: Check if the lake exists in OpenStreetMap at openstreetmap.org
#>   Sites per lake:
#>     ID: 1871997: 2 sites
#>     Blue Mountain Lake: 3 sites
#>     ID: 3361008: 1 sites
#>     Raquette Lake: 3 sites
#> Calculating fetch for multiple lakes...
#>   Buffering sites 10m inward
#>   Angle resolution: 5 degrees
#> Warning in calculate_fetch_multi_lake(sites_with_lakes, all_lakes, utm_epsg, :
#> 3 site(s) not matched to any lake boundary (fetch values will be NA):
#> Raquette_3, Tupper_2, Tupper_3. These lakes may be too small for OpenStreetMap.
#> You can supply your own boundary file using: get_lake_boundary(sites, file =
#> 'your_boundary.gpkg')
#> Processing 4 lake(s)...
#> Using parallel processing with 4 cores for 4 lakes
#> Fetch calculation complete.
#>   Note: 3 unmatched site(s) included with NA fetch values

# Visualize
plot_fetch_map(results)      # Map colored by exposure class
```

![](https://i.imgur.com/pb0B12Z.png)

``` r
plot_fetch_bars(results)     # Bar chart of effective fetch
#> Warning: Removed 3 rows containing missing values or values outside the scale range
#> (`geom_bar()`).
```

![](https://i.imgur.com/aZSrJgm.png)

``` r
plot_fetch_rose(results, 1)  # Rose diagram for the first site
#> Error in `plot_fetch_rose()`:
#> ! Site '1' not found in results
```

<summary>

Session info
</summary>

``` r
sessioninfo::session_info()
#> ─ Session info ───────────────────────────────────────────────────────────────
#>  setting  value
#>  version  R version 4.5.3 (2026-03-11)
#>  os       Ubuntu 24.04.4 LTS
#>  system   x86_64, linux-gnu
#>  ui       X11
#>  language en_GB:en
#>  collate  en_GB.UTF-8
#>  ctype    en_GB.UTF-8
#>  tz       Europe/Madrid
#>  date     2026-04-21
#>  pandoc   3.6.3 @ /usr/lib/rstudio/resources/app/bin/quarto/bin/tools/x86_64/ (via rmarkdown)
#>  quarto   1.8.25 @ /usr/lib/rstudio/resources/app/bin/quarto/bin/quarto
#>
#> ─ Packages ───────────────────────────────────────────────────────────────────
#>  package      * version date (UTC) lib source
#>  class          7.3-23  2025-01-01 [3] RSPM (R 4.4.0)
#>  classInt       0.4-11  2025-01-08 [3] RSPM (R 4.4.0)
#>  cli            3.6.6   2026-04-09 [3] CRAN (R 4.5.3)
#>  curl           7.0.0   2025-08-19 [3] RSPM (R 4.5.0)
#>  DBI            1.3.0   2026-02-25 [3] CRAN (R 4.5.2)
#>  digest         0.6.39  2025-11-19 [3] CRAN (R 4.5.2)
#>  dplyr          1.2.1   2026-04-03 [3] CRAN (R 4.5.3)
#>  e1071          1.7-17  2025-12-18 [3] CRAN (R 4.5.2)
#>  evaluate       1.0.5   2025-08-27 [3] RSPM (R 4.5.0)
#>  farver         2.1.2   2024-05-13 [3] RSPM (R 4.4.2)
#>  fastmap        1.2.0   2024-05-15 [3] RSPM (R 4.4.2)
#>  fs             2.1.0   2026-04-18 [3] CRAN (R 4.5.3)
#>  generics       0.1.4   2025-05-09 [3] RSPM (R 4.5.0)
#>  ggplot2        4.0.2   2026-02-03 [3] CRAN (R 4.5.2)
#>  glue           1.8.1   2026-04-17 [3] CRAN (R 4.5.3)
#>  gtable         0.3.6   2024-10-25 [3] RSPM (R 4.4.0)
#>  htmltools      0.5.9   2025-12-04 [3] CRAN (R 4.5.2)
#>  httr2          1.2.2   2025-12-08 [3] CRAN (R 4.5.2)
#>  KernSmooth     2.23-26 2025-01-01 [3] RSPM (R 4.4.0)
#>  knitr          1.51    2025-12-20 [3] CRAN (R 4.5.2)
#>  labeling       0.4.3   2023-08-29 [3] RSPM (R 4.4.0)
#>  lakefetch    * 0.1.3   2026-04-21 [1] local
#>  lifecycle      1.0.5   2026-01-08 [3] CRAN (R 4.5.2)
#>  magrittr       2.0.5   2026-04-04 [3] CRAN (R 4.5.3)
#>  osmdata        0.3.0   2025-08-23 [3] RSPM (R 4.5.0)
#>  otel           0.2.0   2025-08-29 [3] RSPM (R 4.5.0)
#>  pillar         1.11.1  2025-09-17 [3] RSPM (R 4.5.0)
#>  pkgconfig      2.0.3   2019-09-22 [3] CRAN (R 4.0.1)
#>  proxy          0.4-29  2025-12-29 [3] CRAN (R 4.5.2)
#>  R6             2.6.1   2025-02-15 [3] RSPM (R 4.4.0)
#>  rappdirs       0.3.4   2026-01-17 [3] CRAN (R 4.5.2)
#>  RColorBrewer   1.1-3   2022-04-03 [3] RSPM (R 4.4.0)
#>  Rcpp           1.1.1   2026-01-10 [3] CRAN (R 4.5.2)
#>  reprex         2.1.1   2024-07-06 [3] RSPM (R 4.4.0)
#>  rlang          1.2.0   2026-04-06 [3] CRAN (R 4.5.3)
#>  rmarkdown      2.31    2026-03-26 [3] CRAN (R 4.5.3)
#>  rstudioapi     0.18.0  2026-01-16 [3] CRAN (R 4.5.2)
#>  S7             0.2.1   2025-11-14 [3] CRAN (R 4.5.2)
#>  scales         1.4.0   2025-04-24 [3] RSPM (R 4.4.0)
#>  sessioninfo    1.2.3   2025-02-05 [3] RSPM (R 4.4.0)
#>  sf             1.1-0   2026-02-24 [3] CRAN (R 4.5.2)
#>  tibble         3.3.1   2026-01-11 [3] CRAN (R 4.5.2)
#>  tidyselect     1.2.1   2024-03-11 [3] RSPM (R 4.4.0)
#>  units          1.0-1   2026-03-11 [3] CRAN (R 4.5.2)
#>  vctrs          0.7.3   2026-04-11 [3] CRAN (R 4.5.3)
#>  withr          3.0.2   2024-10-28 [3] RSPM (R 4.4.0)
#>  xfun           0.57    2026-03-20 [3] CRAN (R 4.5.3)
#>  xml2           1.5.2   2026-01-17 [3] CRAN (R 4.5.2)
#>  yaml           2.3.12  2025-12-10 [3] CRAN (R 4.5.2)
#>
#>  [1] /home/frs/R/x86_64-pc-linux-gnu-library/4.0
#>  [2] /usr/local/lib/R/site-library
#>  [3] /usr/lib/R/site-library
#>  [4] /usr/lib/R/library
#>  * ── Packages attached to the search path.
#>
#> ──────────────────────────────────────────────────────────────────────────────
```
