### 2026-04-21 #762 @Pakillo
Also, the shiny app looks great and pretty useful! But it seems that the fetch rose diagram doesn't update when clicking on a new custom point? Please check. Thanks
