### 2026-04-21 #762 @Pakillo
As a suggestion, it might be useful to show the full citation output in the [README](https://github.com/jeremylfarrell/lakefetch#citation) so that users can directly copy and paste it elsewhere?
