### 2026-04-21 #762 @Pakillo
Overall, test coverage is fine, but it might be good to improve coverage for some parts (in `lake_sources.R` and `nhd_integration`) in the future. Results of `devtools::test_coverage()`
