### 2026-04-22 4b4859b
Fix plot_fetch_rose integer index and silent 50km fetch bug
- plot_fetch_rose() now accepts an integer row index (e.g., 1) in
  addition to a site name string; error message now lists available
  sites when a name is not found
- get_highres_fetch() now returns NA for all directions when the site
  point is outside the lake polygon, instead of silently returning
  max_fetch_m (50 km); this fixes misleadingly large fetch values
  reported when GPS/OSM coordinates don't perfectly align

Co-Authored-By: Claude Sonnet 4.6 <noreply@anthropic.com>
