### 2026-04-22 2d62718
Fix Shiny rose diagram for custom points; expand README citation
- Both fetch_app() and fetch_app_upload(): the rose diagram now updates
  correctly when clicking a new custom point on the map. Root cause was
  that the marker-click observer overwrote output$click_results with a
  new renderUI binding, after which the server-level reactive no longer
  drove that output. Fix: custom map-click handler now explicitly
  rebinds output$click_results (same pattern as marker-click), and also
  generates a rose diagram from the click's fetch distances
- README: expand citation section to show the form
[…]
