### 2026-04-24 #762 @jeremylfarrell
Hi Francisco,
Thanks for the thorough editor checks. I can already feel the package maturing! I think all of the issues raised in our exchange are now fixed on the ropensci-software-review branch:

pkgdown website - now live at https://jeremylfarrell.github.io/lakefetch/ with a full reference index grouped by topic, vignettes, and changelog.

Long-running examples - all \donttest{} blocks have been converted to @examplesIf interactive() throughout the package. This uses the roxygen2-native approach and prevents CRAN from running internet-dependent examples while keeping them accessible to users in interactive sessions.

plot_fetch_rose() integer index - the function previously required an exact site name string, so plot_fetch_rose(results, 1) threw an uninformative error. It now accepts either a site name or an integer row index. The error message when a name isn't found now also lists the available site names.

Silent 50 km fetch values - when a site point ended up slightly outside the lake polygon due to GPS/OSM coordinate mismatch, the ray-casting function silently returned max_fetch_m (50,000 m) instead of flagging the problem. The function now checks upfront whether the site is inside the polygon and returns NA for all directions if not, consistent with the existing handling for unmatched sites.

Shiny rose diagram on custom point click - the rose diagram was not updating when clicking a new custom point after having previously clicked a site marker. The root cause was that the marker-click observer overwrote the output$click_results binding, breaking the custom click handler's ability to update the sidebar. Both fetch_app() and fetch_app_upload() are now fixed, and custom point clicks also render a rose diagram.

README citation - the formatted text citation and BibTeX entry are now shown directly in the README for easy copy-paste, alongside the citation("lakefetch") call.

Test coverage - this is tricky. The low coverage in lake_sources.R and nhd_integration.R reflects the network-dependent OSM and NHD code paths, which are difficult to test without deep mocking. Ultimately I think if a user does not have network access then this package might not be for them. That said, if you think it is important I am happy to explore it.
