### 2026-05-03 #762 @maurolepore
I see @Pakillo has completed editor checks and is searching for reviewers. Looking forward to seeing this move through review.

Thanks for sharing your work with the rOpenSci community.
