### 2026-05-12 3282b32
Add rOpenSci software peer review badge to README
Co-Authored-By: Claude Sonnet 4.6 <noreply@anthropic.com>
