### 2026-05-12 296b8e8
Deploying to gh-pages from @ jeremylfarrell/lakefetch@3282b3221b15b24d5fc433cc0c888d22ec5a5771 🚀
