+++ index.md
@@ -39,12 +39,14 @@ system representations.
 Install from CRAN:
 
 ``` r
+
 install.packages("lakefetch")
 ```
 
 Or install the development version from GitHub:
 
 ``` r
+
 # install.packages("remotes")
 remotes::install_github("jeremylfarrell/lakefetch")
 ```
@@ -52,6 +54,7 @@ remotes::install_github("jeremylfarrell/lakefetch")
 ## Quick Start
 
 ``` r
+
 library(lakefetch)
 
 # Load sampling sites from a CSV with latitude/longitude columns
@@ -82,13 +85,13 @@ results$results
 
 For each site, lakefetch returns:
 
-| Metric                  | Description                                                     |
-|-------------------------|-----------------------------------------------------------------|
-| `fetch_0` … `fetch_355` | Distance to shore (m) at each compass bearing                   |
-| `fetch_mean`            | Mean of all directional fetches                                 |
-| `fetch_max`             | Maximum directional fetch (longest open water distance)         |
-| `fetch_effective`       | Effective fetch (default: mean of 3 highest directional values) |
-| `exposure_class`        | Categorical exposure classification                             |
+| Metric | Description |
+|----|----|
+| `fetch_0` … `fetch_355` | Distance to shore (m) at each compass bearing |
+| `fetch_mean` | Mean of all directional fetches |
+| `fetch_max` | Maximum directional fetch (longest open water distance) |
+| `fetch_effective` | Effective fetch (default: mean of 3 highest directional values) |
+| `exposure_class` | Categorical exposure classification |
 
 Three effective fetch methods are available: `"top3"` (default),
 `"max"`, and `"cosine"` (the SPM/CERC cosine-weighted method from the
@@ -126,6 +129,7 @@ Shore Protection Manual).
 ## Example Workflow
 
 ``` r
+
 library(lakefetch)
 
 # Built-in example data
@@ -148,25 +152,26 @@ fetch_app(results)
 
 ## Key Functions
 
-| Function                                                                                               | Description                                                        |
-|--------------------------------------------------------------------------------------------------------|--------------------------------------------------------------------|
-| [`load_sites()`](https://jeremylfarrell.github.io/lakefetch/reference/load_sites.md)                   | Load and validate sampling sites from CSV or data frame            |
-| [`get_lake_boundary()`](https://jeremylfarrell.github.io/lakefetch/reference/get_lake_boundary.md)     | Download lake polygons from OSM or load from local file            |
-| [`fetch_calculate()`](https://jeremylfarrell.github.io/lakefetch/reference/fetch_calculate.md)         | Calculate directional fetch, effective fetch, and exposure class   |
-| [`add_lake_depth()`](https://jeremylfarrell.github.io/lakefetch/reference/add_lake_depth.md)           | Estimate lake depth from surface area (Cael et al. 2017)           |
-| [`add_lake_context()`](https://jeremylfarrell.github.io/lakefetch/reference/add_lake_context.md)       | Add NHD hydrological context (US lakes; requires nhdplusTools)     |
+| Function | Description |
+|----|----|
+| [`load_sites()`](https://jeremylfarrell.github.io/lakefetch/reference/load_sites.md) | Load and validate sampling sites from CSV or data frame |
+| [`get_lake_boundary()`](https://jeremylfarrell.github.io/lakefetch/reference/get_lake_boundary.md) | Download lake polygons from OSM or load from local file |
+| [`fetch_calculate()`](https://jeremylfarrell.github.io/lakefetch/reference/fetch_calculate.md) | Calculate directional fetch, effective fetch, and exposure class |
+| [`add_lake_depth()`](https://jeremylfarrell.github.io/lakefetch/reference/add_lake_depth.md) | Estimate lake depth from surface area (Cael et al. 2017) |
[…]
