+++ pkgdown.yml
@@ -1,10 +1,10 @@
-pandoc: 3.1.11
+pandoc: 3.8.3
 pkgdown: 2.2.0
 pkgdown_sha: ~
 articles:
   getting-started: getting-started.html
   validation: validation.html
-last_built: 2026-04-22T00:22Z
+last_built: 2026-05-12T01:09Z
 urls:
   reference: https://jeremylfarrell.github.io/lakefetch/reference
   article: https://jeremylfarrell.github.io/lakefetch/articles
