### 2026-05-12 #762 @jeremylfarrell
Thanks @Pakillo  and @maurolepore
The end of the semester got away from me but i have now added the news and badge
Looking forward to the review!
