### 2026-06-01 dd078fa
Address rOpenSci review (#762): v0.1.4
Response to peer review by Jorrit Mesman. All seven review issues addressed
plus two bonus fixes found while testing.

Main issues:
- Sample sites in inst/extdata/sample_sites.csv and adirondack_sites
  dataset moved inside their lake polygons; fetch_calculate() now returns
  non-NA exposure for all example sites.
- get_lake_boundary() gains timeout (default 90s) and simplify_tolerance_m
  (default 0, in meters via st_simplify) arguments for very large or
  complex lakes such as Malaren, Vattern, Vortsjarv. simplify_tolerance_m
  works for both OSM and fi
[…]
