+++ tests/testthat/test-lake-sources-helpers.R
@@ -450,3 +450,157 @@ test_that("load_lake_file handles multipolygon geometry", {
 
   unlink(tmp_gpkg)
 })
+
+# --- get_lake_boundary new arguments: simplify_tolerance_m, timeout ---
+
+test_that("simplify_tolerance_m = 0 leaves polygon vertex count unchanged", {
+  # Create a polygon with many vertices so simplification has something to do
+  n <- 60
+  theta <- seq(0, 2 * pi, length.out = n + 1)[-(n + 1)]
+  cx <- -74; cy <- 43; r <- 0.05
+  # Add small perturbation so the ring isn't a perfect circle
+  set.seed(42)
+  ring_x <- cx + r * cos(theta) + rnorm(n, sd = 0.0005)
+  ring_y <- cy + r * sin(theta) + rnorm(n, sd = 0.0005)
+  ring <- cbind(c(ring_x, ring_x[1]), c(ring_y, ring_y[1]))
+
+  lake_wgs84 <- sf::st_sf(
+    name = "TestLake",
+    geometry = sf::st_sfc(sf::st_polygon(list(ring)), crs = 4326)
+  )
+
+  tmp_gpkg <- tempfile(fileext = ".gpkg")
+  sf::st_write(lake_wgs84, tmp_gpkg, quiet = TRUE)
+
+  site_sf <- sf::st_sf(
+    Site = "Test",
+    geometry = sf::st_sfc(sf::st_point(c(cx, cy)), crs = 4326)
+  )
+
+  raw <- lakefetch::get_lake_boundary(site_sf, file = tmp_gpkg,
+                                       simplify_tolerance_m = 0)
+  v_raw <- length(unlist(sf::st_geometry(raw$all_lakes))) / 2
+  expect_equal(v_raw, n + 1)  # ring close-point makes it n + 1 coords
+
+  unlink(tmp_gpkg)
+})
+
+test_that("simplify_tolerance_m > 0 reduces polygon vertex count", {
+  n <- 60
+  theta <- seq(0, 2 * pi, length.out = n + 1)[-(n + 1)]
+  cx <- -74; cy <- 43; r <- 0.05
+  set.seed(42)
+  ring_x <- cx + r * cos(theta) + rnorm(n, sd = 0.0005)
+  ring_y <- cy + r * sin(theta) + rnorm(n, sd = 0.0005)
+  ring <- cbind(c(ring_x, ring_x[1]), c(ring_y, ring_y[1]))
+
+  lake_wgs84 <- sf::st_sf(
+    name = "TestLake",
+    geometry = sf::st_sfc(sf::st_polygon(list(ring)), crs = 4326)
+  )
+
+  tmp_gpkg <- tempfile(fileext = ".gpkg")
+  sf::st_write(lake_wgs84, tmp_gpkg, quiet = TRUE)
+
+  site_sf <- sf::st_sf(
+    Site = "Test",
+    geometry = sf::st_sfc(sf::st_point(c(cx, cy)), crs = 4326)
+  )
+
+  raw  <- lakefetch::get_lake_boundary(site_sf, file = tmp_gpkg,
+                                        simplify_tolerance_m = 0)
+  simp <- lakefetch::get_lake_boundary(site_sf, file = tmp_gpkg,
+                                        simplify_tolerance_m = 500)
+
+  v_raw  <- length(unlist(sf::st_geometry(raw$all_lakes))) / 2
+  v_simp <- length(unlist(sf::st_geometry(simp$all_lakes))) / 2
+
+  expect_lt(v_simp, v_raw)
+  expect_gt(v_simp, 3)  # Still a valid polygon
+
+  # Simplified polygon must still be valid sf geometry
[…]
