+++ R/lake_sources.R
@@ -8,6 +8,15 @@
 #'
 #' @param sites A data.frame with latitude and longitude columns, or an sf object.
 #' @param file Optional file path to a shapefile or geopackage with lake boundaries.
+#' @param timeout Integer; Overpass API query timeout in seconds. Default is 90.
+#'   Increase for very large lakes (e.g., \code{timeout = 300} for Mälaren or
+#'   Great Lakes) or when server load is high.
+#' @param simplify_tolerance_m Numeric; if greater than 0, simplify lake
+#'   polygons with \code{sf::st_simplify(dTolerance = simplify_tolerance_m)}
+#'   (in meters, applied in the UTM projection). Useful for very large or
+#'   complex lakes where an exact coastline is not needed and a coarser
+#'   polygon greatly speeds up fetch ray-casting. Typical values: 50-500 m
+#'   for large lakes (e.g., Mälaren, Vättern). Default is 0 (no simplification).
 #'
 #' @return A list with elements:
 #'   \item{all_lakes}{sf object with lake polygons in UTM projection}
@@ -19,18 +28,58 @@
 #' Otherwise, the function downloads lake boundaries from OpenStreetMap
 #' based on the bounding box of the provided sites.
 #'
+#' For very large lakes (> ~500 km\eqn{^2}), the default 90-second Overpass
+#' API timeout may be exceeded. Use \code{timeout = 300} or higher in those
+#' cases. For lakes with very complex shorelines (e.g., Mälaren, Vättern,
+#' Võrtsjärv), additionally pass \code{simplify_tolerance_m = 100} (or higher)
+#' to coarsen the polygon and speed up downstream fetch calculations.
+#'
 #' @examplesIf interactive()
 #' csv_path <- system.file("extdata", "sample_sites.csv", package = "lakefetch")
 #' sites <- load_sites(csv_path)
 #' lake_data <- get_lake_boundary(sites)
 #'
+#' # For very large lakes, increase the timeout
+#' lake_data <- get_lake_boundary(sites, timeout = 300)
+#'
+#' # For large/complex lakes, also coarsen the shoreline
+#' lake_data <- get_lake_boundary(sites, timeout = 300,
+#'                                simplify_tolerance_m = 100)
+#'
 #' @export
-get_lake_boundary <- function(sites, file = NULL) {
-  if (!is.null(file)) {
-    return(load_lake_file(sites, file))
+get_lake_boundary <- function(sites, file = NULL, timeout = 90,
+                              simplify_tolerance_m = 0) {
+  result <- if (!is.null(file)) {
+    load_lake_file(sites, file)
   } else {
-    return(download_lake_osm(sites))
+    download_lake_osm(sites, timeout = timeout)
   }
+
+  if (isTRUE(simplify_tolerance_m > 0) &&
+      !is.null(result$all_lakes) && nrow(result$all_lakes) > 0) {
+    message("Simplifying lake polygons (dTolerance = ",
+            simplify_tolerance_m, " m)...")
+    n_before <- sum(vapply(sf::st_geometry(result$all_lakes),
+                           function(g) length(unlist(g)), integer(1))) / 2
+    simplified <- sf::st_simplify(result$all_lakes,
+                                   dTolerance = simplify_tolerance_m,
+                                   preserveTopology = TRUE)
+    # Repair any geometries that simplification rendered invalid/empty
+    not_empty <- !sf::st_is_empty(simplified)
+    simplified <- simplified[not_empty, ]
+    bad <- !sf::st_is_valid(simplified)
+    if (any(bad)) {
+      simplified[bad, ] <- sf::st_make_valid(simplified[bad, ])
+    }
[…]
