### 2026-06-01 f172a30
Deploying to gh-pages from @ jeremylfarrell/lakefetch@dd078fa37a5fe8e9f4306ea8a80419e2127f1890 🚀
