+++ pkgdown.yml
@@ -4,7 +4,7 @@ pkgdown_sha: ~
 articles:
   getting-started: getting-started.html
   validation: validation.html
-last_built: 2026-05-12T01:09Z
+last_built: 2026-06-01T18:53Z
 urls:
   reference: https://jeremylfarrell.github.io/lakefetch/reference
   article: https://jeremylfarrell.github.io/lakefetch/articles

[…]
