### 2026-06-30 faf871e
Address rOpenSci review (#762, Khondula): v0.1.5
Response to second peer review by Khondula. All eleven review issues plus
a couple of incidental fixes are addressed.

Bug fixes:
- download_lake_osm() previously computed the UTM zone from raw input
  coordinates, producing nonsense EPSG codes such as 32683364 when given
  an sf object already in a projected CRS (e.g., example_lake in UTM 18N).
  The zone is now derived from the WGS84-transformed centroid.
- assign_sites_to_lakes() now emits a warning() in addition to the
  diagnostic messages when sites cannot be matched. Previously the
  mess
[…]
