+++ tests/testthat/test-lake-sources-mocked.R
@@ -227,3 +227,175 @@ test_that("get_lake_boundary dispatches to OSM download when no file", {
   expect_type(result, "list")
   expect_true("all_lakes" %in% names(result))
 })
+
+# --- v0.1.5 regression tests for Khondula review fixes ---
+
+test_that("download_lake_osm computes valid UTM EPSG for sf input in projected CRS", {
+  # Khondula: passing example_lake (EPSG:32618 UTM) to get_lake_boundary
+  # produced EPSG:32683364 because the UTM zone was derived from raw projected
+  # coordinates instead of WGS84.
+  sites_wgs <- sf::st_sf(
+    Site = "S1",
+    geometry = sf::st_sfc(sf::st_point(c(-74.05, 43.05)), crs = 4326)
+  )
+  sites_utm <- sf::st_transform(sites_wgs, 32618)
+
+  mock_polys <- create_mock_water_polys()
+  mock_polys$area_km2 <- as.numeric(sf::st_area(
+    sf::st_transform(mock_polys, 32618))) / 1e6
+
+  local_mocked_bindings(
+    download_lake_osm_single = function(bbox_vec, lake_names = NULL,
+                                         overpass_servers = NULL,
+                                         name_only = FALSE, ...) {
+      list(mock_polys)
+    },
+    .package = "lakefetch"
+  )
+
+  result <- lakefetch:::download_lake_osm(sites_utm)
+
+  # Valid northern-hemisphere UTM EPSGs are 32601-32660. The pre-fix bug
+  # produced numbers like 32683364.
+  expect_gte(result$utm_epsg, 32601)
+  expect_lte(result$utm_epsg, 32660)
+  # The expected zone for (-74°, 43°) is 18N -> EPSG:32618
+  expect_equal(result$utm_epsg, 32618)
+})
+
+test_that("download_lake_osm detects flexible lat/lon column names", {
+  # Khondula: load_sites accepts 'lat'/'lon', get_lake_boundary did not.
+  # Now get_lake_boundary should accept abbreviated, capitalized, and y/x
+  # column names on data.frame input.
+  mock_polys <- create_mock_water_polys()
+  mock_polys$area_km2 <- as.numeric(sf::st_area(
+    sf::st_transform(mock_polys, 32618))) / 1e6
+
+  local_mocked_bindings(
+    download_lake_osm_single = function(bbox_vec, lake_names = NULL,
+                                         overpass_servers = NULL,
+                                         name_only = FALSE, ...) {
+      list(mock_polys)
+    },
+    .package = "lakefetch"
+  )
+
+  # Variant 1: 'lat'/'lon'
+  sites_lat <- data.frame(Site = "S1", lat = 43.005, lon = -74.005,
+                          stringsAsFactors = FALSE)
+  expect_no_error(
+    result1 <- lakefetch:::download_lake_osm(sites_lat)
+  )
+  expect_true("utm_epsg" %in% names(result1))
+
+  # Variant 2: capitalized
+  sites_cap <- data.frame(Site = "S1", Latitude = 43.005, Longitude = -74.005,
+                          stringsAsFactors = FALSE)
+  expect_no_error(lakefetch:::download_lake_osm(sites_cap))
+
+  # Variant 3: y/x convention
+  sites_xy <- data.frame(Site = "S1", y = 43.005, x = -74.005,
+                         stringsAsFactors = FALSE)
+  expect_no_error(lakefetch:::download_lake_osm(sites_xy))
+
+  # Variant 4: missing both columns -> actionable error listing available cols
+  sites_bad <- data.frame(Site = "S1", foo = 1, bar = 2,
+                          stringsAsFactors = FALSE)
[…]
