+++ R/lake_sources.R
@@ -328,12 +328,27 @@ download_lake_osm <- function(sites_df, timeout = 90) {
   sf::sf_use_s2(FALSE)
   on.exit(sf::sf_use_s2(TRUE), add = TRUE)
 
-  # Convert to sf object (WGS84)
+  # Convert to sf object (WGS84). For data.frame input, detect lat/lon
+  # columns flexibly (case-insensitive, accepts "lat"/"latitude", "lon"/"long"
+  # /"longitude") so users don't have to pre-process with load_sites() if
+  # their data already has standard-ish column names.
   if (inherits(sites_df, "sf")) {
     sites_sf <- sf::st_transform(sites_df, 4326)
   } else {
+    col_lower <- tolower(names(sites_df))
+    lat_idx <- which(col_lower %in% c("latitude", "lat", "y"))[1]
+    if (is.na(lat_idx)) lat_idx <- grep("^lat", col_lower)[1]
+    lon_idx <- which(col_lower %in% c("longitude", "lon", "long", "lng", "x"))[1]
+    if (is.na(lon_idx)) lon_idx <- grep("^lon", col_lower)[1]
+    if (is.na(lat_idx) || is.na(lon_idx)) {
+      stop("Could not find latitude / longitude columns in input data.frame.\n",
+           "  Available columns: ", paste(names(sites_df), collapse = ", "), "\n",
+           "  Either rename columns to 'latitude' / 'longitude' or pass the\n",
+           "  data through load_sites() first.", call. = FALSE)
+    }
     sites_sf <- sf::st_as_sf(sites_df,
-                              coords = c("longitude", "latitude"),
+                              coords = c(names(sites_df)[lon_idx],
+                                         names(sites_df)[lat_idx]),
                               crs = 4326)
   }
 
@@ -383,7 +398,34 @@ download_lake_osm <- function(sites_df, timeout = 90) {
     water_list <- download_lake_osm_single(bbox_vec, cluster_lake_names,
                                            overpass_servers, timeout = timeout)
   } else {
-    # --- Large spread: per-cluster small bbox queries ---
+    # --- Large spread ---
+    # Fast path: if every site has a known lake name, issue a single
+    # name-filtered Overpass query covering the whole site bbox. The name
+    # regex is selective enough that returns stay small, and one query is
+    # dramatically faster than per-cluster broad queries.
+    all_named <- !is.null(lake_name_col) &&
+      all(!is.na(sites_sf[[lake_name_col]]) &
+          nchar(trimws(as.character(sites_sf[[lake_name_col]]))) > 0)
+
+    if (isTRUE(all_named)) {
+      unique_names <- unique(sites_sf[[lake_name_col]])
+      message("  Site spread is ", round(site_spread, 1),
+              " degrees with all sites named - using single name-filtered query for ",
+              length(unique_names), " lake(s)")
+      bbox_vec_full <- c(bbox["xmin"], bbox["ymin"],
+                         bbox["xmax"], bbox["ymax"])
+      names(bbox_vec_full) <- c("left", "bottom", "right", "top")
+      name_res <- query_osm_by_name(bbox_vec_full, unique_names,
+                                     overpass_servers,
+                                     timeout = max(timeout, 120))
+      if (!is.null(name_res)) {
+        water_list <- c(water_list, extract_osm_polys(name_res))
+      }
+      tryCatch(osmdata::set_overpass_url(overpass_servers[1]),
+               error = function(e) NULL)
+    } else {
+
+    # --- Per-cluster small bbox queries (used when lake names are missing) ---
[…]
