+++ news/index.md
@@ -1,5 +1,85 @@
 # Changelog
 
+## lakefetch 0.1.5
+
+Changes in response to the second rOpenSci peer review by Khondula
+(ropensci/software-review#762).
+
+### Bug fixes
+
+- **Invalid UTM EPSG when sf input is in a projected CRS**: When a user
+  passed an sf object in a non-WGS84 CRS (e.g., the bundled
+  `example_lake` in UTM zone 18N) to
+  [`get_lake_boundary()`](https://jeremylfarrell.github.io/lakefetch/reference/get_lake_boundary.md),
+  the UTM zone detection was performed on the raw projected coordinates
+  and produced nonsensical EPSG codes such as `EPSG:32683364`. The UTM
+  zone is now computed from the centroid of the WGS84-transformed sites
+  instead of the raw input.
+- **Sites unassigned to lakes now emit a warning, not just a message**:
+  When
+  [`assign_sites_to_lakes()`](https://jeremylfarrell.github.io/lakefetch/reference/assign_sites_to_lakes.md)
+  cannot match one or more sites, it now emits a
+  [`warning()`](https://rdrr.io/r/base/warning.html) in addition to the
+  diagnostic messages. The previous message-only behavior could be
+  missed in console output, leading users to believe the workflow had
+  fully succeeded when in fact some sites had NA fetch values.
+- **Getting-started vignette runnable end-to-end**: The Quick Start
+  chunks were reorganized so the `sites` object is not overwritten with
+  mismatched coordinates, and the network-dependent chunks are
+  explicitly marked `eval = FALSE` with a note explaining how to run
+  them interactively.
+
+### Improvements
+
+- **Flexible column-name detection for
+  [`get_lake_boundary()`](https://jeremylfarrell.github.io/lakefetch/reference/get_lake_boundary.md)**:
+  The function now auto-detects latitude / longitude columns on
+  data.frame input the same way
+  [`load_sites()`](https://jeremylfarrell.github.io/lakefetch/reference/load_sites.md)
+  does (accepts “lat” / “latitude” / “y” and “lon” / “long” /
+  “longitude” / “lng” / “x”, case-insensitive). Previously it required
+  the exact column names produced by
+  [`load_sites()`](https://jeremylfarrell.github.io/lakefetch/reference/load_sites.md).
+- **Faster downloads for spread-out named lakes**: When all sites have a
+  known lake name and the site spread exceeds 0.5 degrees,
+  [`get_lake_boundary()`](https://jeremylfarrell.github.io/lakefetch/reference/get_lake_boundary.md)
+  now issues a single name-filtered Overpass query covering the union
+  bounding box instead of looping through per-cluster broad queries. For
+  datasets like `wisconsin_lakes` (3 lakes spread across ~1 degree),
+  this typically reduces download time from many minutes to under a
+  minute.
+- **[`add_weather_context()`](https://jeremylfarrell.github.io/lakefetch/reference/add_weather_context.md)
+  docs expanded**: The function documentation now describes what the
+  function does (queries Open-Meteo, combines with fetch, integrates
+  wave energy), and the `windows_hours` argument is documented
+  explicitly (per-window column names, intended use as look-back windows
+  ending at each sample’s `datetime`).
+- **[`assign_sites_to_lakes()`](https://jeremylfarrell.github.io/lakefetch/reference/assign_sites_to_lakes.md)
+  `tolerance_m` documented**: The default (50 m via
+  [`lakefetch_options()`](https://jeremylfarrell.github.io/lakefetch/reference/lakefetch_options.md))
+  is now explained, with guidance to increase it (200-500 m) for sites
+  with appreciable GPS error or near-shore coordinates. The example uses
[…]
