+++ pkgdown.yml
@@ -4,7 +4,7 @@ pkgdown_sha: ~
 articles:
   getting-started: getting-started.html
   validation: validation.html
-last_built: 2026-06-01T18:53Z
+last_built: 2026-06-30T02:25Z
 urls:
   reference: https://jeremylfarrell.github.io/lakefetch/reference
   article: https://jeremylfarrell.github.io/lakefetch/articles

[…]
