### 2026-07-01 874e717
Address rOpenSci review (#762, Khondula): v0.1.6
Small follow-up fixes after Khondula's positive re-test of v0.1.5.

Bug fixes:
- plot_fetch_rose() now resets par(new = FALSE) before drawing. Without
  this, running the function immediately after fetch_app() (which leaves
  the device with par$new = TRUE) produced a rose that overlaid the
  previous plot or lacked a background.

Improvements:
- Rose diagram color switched from blue (rgb(0.2, 0.5, 0.8, 0.4) with
  steelblue border) to semi-transparent purple with a #7B3E9E border.
  Purple matches the common wind-rose convention and avoids visu
[…]
