### 2026-07-01 2de2645
Deploying to gh-pages from @ jeremylfarrell/lakefetch@874e717bdd0cd8c532f9699b1d121becb9338bd8 🚀
