### 2026-07-01 7f2c1cf
Address rOpenSci review (#762, Khondula): v0.1.7
Small follow-up after Khondula's re-test of v0.1.6.

Bug fixes:
- plot_fetch_rose() real fix for overplotting after fetch_app() /
  plot_fetch_map() / plot_fetch_bars(). The v0.1.6 par(new = FALSE) reset
  was insufficient: ggplot2 (used by the map / bars plots and internally
  by the Shiny app) leaves an active grid viewport on the graphics
  device, and base R plot.new() cannot clear a grid viewport. The
  function now calls grid::grid.newpage() at entry so subsequent base R
  plotting draws into a clean region.
- wisconsin_lakes Geneva Lake c
[…]
