+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: lakefetch
 Title: Calculate Fetch and Wave Exposure for Lake Sampling Points
-Version: 0.1.6
+Version: 0.1.7
 Authors@R:
     person(given = "Jeremy",
            family = "Lynch Farrell",
