### 2026-07-01 24ddddc
Deploying to gh-pages from @ jeremylfarrell/lakefetch@7f2c1cfd9ac5ea6d2b0862f1422a5f0a674e264a 🚀
