### 2026-07-02 f5b5449
Address rOpenSci review (#762, Khondula): v0.1.8
Follow-up to v0.1.7 after Khondula confirmed the rose-overplotting fix
still did not resolve the issue on her system (likely macOS quartz
device behavior).

Bug fixes:
- plot_fetch_rose() now closes the current interactive graphics device
  (RStudioGD / windows / X11 / quartz) on entry if one is open, then
  lets the next plotting call reopen a fresh device. This matches the
  reviewer's own confirmed manual workaround (calling dev.off() before
  the rose plot). The v0.1.7 grid::grid.newpage() approach did not
  clear the device state on all pla
[…]
