### 2026-07-02 e1e3a68
Deploying to gh-pages from @ jeremylfarrell/lakefetch@f5b5449d1f53185e3e69cf85e1934f3388d29003 🚀
