### 2026-07-10 7a82d5c
Address rOpenSci editor issue (#762, Pakillo): v0.1.9
Response to issue lakefetch#2 opened by rOpenSci editor Pakillo during
final review checks.

Bug fixes:
- get_lake_boundary() worst-case wall time is now bounded to about five
  minutes. Previous versions could sit for 20+ minutes on a throttled or
  broken Overpass session because each of three query types retried
  three times, and each retry triggered another 60-second rate-limit
  backoff inside osmdata. Two changes cap the wait:
    * Errors that are not usefully retryable now short-circuit after the
      first attempt: the "arguments
[…]
