### 2026-07-10 15d66eb
Deploying to gh-pages from @ jeremylfarrell/lakefetch@7a82d5c04ca2f3da383cd2c61cee7275234d2edb 🚀
