### 2026-07-13 2493a69
Address rOpenSci editor review (#762, Pakillo): v0.1.10
Response to Pakillo's final review comments on the rOpenSci software
review thread (ropensci/software-review#762 items E1-E7).

E5 - Global wall-clock cap on get_lake_boundary():
- New total_timeout_s argument (default 300 seconds) bounds aggregate
  wall time across every query type and per-cluster loop, not just the
  single-Overpass-call timeout. The function aborts and returns
  whatever it has collected so far when the budget is exceeded.

E4 - GeoPackage export from fetch_app_upload() now writes three layers:
- sites (points with fe
[…]
