+++ man/get_lake_boundary.Rd
@@ -4,7 +4,13 @@
 \alias{get_lake_boundary}
 \title{Get Lake Boundary}
 \usage{
-get_lake_boundary(sites, file = NULL, timeout = 90, simplify_tolerance_m = 0)
+get_lake_boundary(
+  sites,
+  file = NULL,
+  timeout = 90,
+  simplify_tolerance_m = 0,
+  total_timeout_s = 300
+)
 }
 \arguments{
 \item{sites}{A data.frame with latitude and longitude columns, or an sf object.}
@@ -21,6 +27,14 @@ polygons with \code{sf::st_simplify(dTolerance = simplify_tolerance_m)}
 complex lakes where an exact coastline is not needed and a coarser
 polygon greatly speeds up fetch ray-casting. Typical values: 50-500 m
 for large lakes (e.g., Mälaren, Vättern). Default is 0 (no simplification).}
+
+\item{total_timeout_s}{Numeric; hard cap in seconds on the wall-clock time
+\code{get_lake_boundary()} will spend downloading from OSM before it
+aborts and returns whatever partial results it has collected. This is
+independent of the per-query \code{timeout}: a slow Overpass server or
+a large multi-cluster query set can still take much longer than
+\code{timeout} in aggregate. Default 300 (5 minutes), set to \code{Inf}
+to disable.}
 }
 \value{
 A list with elements:
