+++ reference/example_lake.md
@@ -1,8 +1,11 @@
-# Example Circular Lake Polygon
+# Blue Mountain Lake Polygon (Example Lake)
 
-A simple circular lake polygon for demonstration and testing purposes.
-This synthetic lake has known geometry (1 km radius) which makes it
-useful for validating fetch calculations.
+The OpenStreetMap boundary polygon for Blue Mountain Lake in Hamilton
+County, New York. Bundled with the package so that fetch examples can
+run offline (no internet or Overpass API call required) and pkgdown
+pages can render plot output. The coordinates match the sites in
+`system.file("extdata", "sample_sites.csv", package = "lakefetch")`, so
+the two datasets can be used together end-to-end.
 
 ## Usage
 
@@ -12,52 +15,57 @@ example_lake
 
 ## Format
 
-An sf object with 1 row and 3 variables:
+An sf object with 1 row and 3 fields plus geometry:
 
 - osm_id:
 
-  Identifier (synthetic)
+  OSM relation identifier
 
 - name:
 
-  Lake name
+  Lake name ("Blue Mountain Lake")
 
 - area_km2:
 
-  Surface area in square kilometers (~3.14 km²)
+  Surface area in square kilometers
 
 - geometry:
 
-  POLYGON geometry in UTM Zone 18N (EPSG:32618)
+  MULTIPOLYGON geometry in UTM Zone 18N (EPSG:32618)
 
 ## Source
 
-Synthetic data for demonstration and validation
-
-## Details
-
-The lake is centered at UTM coordinates (500000, 4800000) with a radius
-of 1000 meters. For a site at the center, fetch should equal 1000 m in
-all directions.
+Downloaded from OpenStreetMap <https://www.openstreetmap.org/>. See
+`data-raw/create_example_data.R` for the exact query.
 
 ## Examples
 
 ``` r
-# Load the dataset
 data(example_lake)
-
-# View structure
 print(example_lake)
 #> Simple feature collection with 1 feature and 3 fields
 #> Geometry type: POLYGON
 #> Dimension:     XY
-#> Bounding box:  xmin: 499000 ymin: 4799000 xmax: 501000 ymax: 4801000
+#> Bounding box:  xmin: 541565.8 ymin: 4855616 xmax: 545660 ymax: 4857845
 #> Projected CRS: WGS 84 / UTM zone 18N
-#>        osm_id                  name area_km2                       geometry
-#> 1 example_001 Example Circular Lake 3.141593 POLYGON ((501000 4800000, 5...
+#>    osm_id               name area_km2                       geometry
+#> 1 2202972 Blue Mountain Lake 5.055174 POLYGON ((543523.8 4857818,...
 
 # Plot the lake
 library(ggplot2)
 ggplot(example_lake) + geom_sf()
 
+
+# Load matching sample sites (they lie inside this polygon) and
+# compute fetch end-to-end without touching OSM. First convert
+# example_lake into the multi-lake list format that fetch_calculate()
+# expects:
+sites <- load_sites(system.file("extdata", "sample_sites.csv",
+                                 package = "lakefetch"))
+#> Loading data from: /home/runner/work/_temp/Library/lakefetch/extdata/sample_sites.csv
+#>   Loaded 2 rows with columns: Site, latitude, longitude, lake.name
+#>   Using columns: Latitude = latitude, Longitude = longitude
+#>   Preserved lake name column: lake.name
+#>   Final valid samples: 2
+#>   Detected location from column 'lake.name': Blue Mountain Lake
 ```
