### 2026-07-14 adb1951
Address rOpenSci review (#762, Pakillo): v0.1.11
Follow-up to Pakillo's v0.1.10 re-test comments on ropensci/software-review#762.

Bug fixes:
- total_timeout_s is now checked at every Overpass query-type boundary
  in the small-spread path (previously only enforced in the multi-cluster
  loop). A short budget will now abort before all three broad queries
  complete. Documented honestly: the budget remains a soft cap because
  a single osmdata::osmdata_sf() call can itself take many minutes due
  to osmdata's own 60-second-per-retry backoff for HTTP 429 responses,
  which we cannot interrupt sa
[…]
