+++ R/lake_sources.R
@@ -17,13 +17,19 @@
 #'   complex lakes where an exact coastline is not needed and a coarser
 #'   polygon greatly speeds up fetch ray-casting. Typical values: 50-500 m
 #'   for large lakes (e.g., Mälaren, Vättern). Default is 0 (no simplification).
-#' @param total_timeout_s Numeric; hard cap in seconds on the wall-clock time
-#'   \code{get_lake_boundary()} will spend downloading from OSM before it
-#'   aborts and returns whatever partial results it has collected. This is
-#'   independent of the per-query \code{timeout}: a slow Overpass server or
-#'   a large multi-cluster query set can still take much longer than
-#'   \code{timeout} in aggregate. Default 300 (5 minutes), set to \code{Inf}
-#'   to disable.
+#' @param total_timeout_s Numeric; soft wall-clock budget in seconds on the
+#'   total time \code{get_lake_boundary()} will spend downloading from OSM.
+#'   The budget is consulted at natural breakpoints (between Overpass query
+#'   types, between clusters for spread-out sites, and between name-filtered
+#'   queries) and aborts further work when exceeded. It is NOT a hard cap
+#'   on a single \code{osmdata::osmdata_sf()} call: if Overpass returns
+#'   HTTP 429, osmdata does its own 60-second-per-retry backoff loop
+#'   internally and we cannot safely interrupt that from R without risking
+#'   a segfault on Windows. So on a heavily throttled server a single call
+#'   may still exceed \code{total_timeout_s} by several minutes. For a hard
+#'   ceiling, wrap the call in \code{R.utils::withTimeout()} yourself, or
+#'   supply a local boundary file via the \code{file} argument. Default
+#'   300 seconds (5 minutes); set to \code{Inf} to disable.
 #'
 #' @return A list with elements:
 #'   \item{all_lakes}{sf object with lake polygons in UTM projection}
@@ -96,6 +102,48 @@ get_lake_boundary <- function(sites, file = NULL, timeout = 90,
 #' Groups sites by proximity using a grid-based approach. Sites within 0.1
 #' degrees (~11 km) of each other are placed in the same cluster.
 #'
+#' Build an empty lake-boundary result for early-abort paths
+#'
+#' Used when get_lake_boundary() short-circuits (e.g., because the
+#' total_timeout_s wall-clock cap has been reached) so downstream code can
+#' still treat the return value as a normal result and assign NA fetch to
+#' every site.
+#'
+#' @param sites A data.frame with latitude / longitude columns, or an sf.
+#' @return A list with the same shape as download_lake_osm()'s empty branch.
+#' @noRd
+make_empty_osm_result <- function(sites) {
+  if (inherits(sites, "sf")) {
+    sites_sf <- sf::st_transform(sites, 4326)
+  } else {
+    col_lower <- tolower(names(sites))
+    lat_idx <- which(col_lower %in% c("latitude", "lat", "y"))[1]
+    if (is.na(lat_idx)) lat_idx <- grep("^lat", col_lower)[1]
+    lon_idx <- which(col_lower %in% c("longitude", "lon", "long", "lng", "x"))[1]
+    if (is.na(lon_idx)) lon_idx <- grep("^lon", col_lower)[1]
+    sites_sf <- sf::st_as_sf(sites,
+                              coords = c(names(sites)[lon_idx],
+                                          names(sites)[lat_idx]),
+                              crs = 4326)
+  }
+  centroid_ll <- suppressWarnings(
+    sf::st_coordinates(sf::st_centroid(sf::st_union(sites_sf)))
+  )
+  utm_zone <- floor((centroid_ll[1] + 180) / 6) + 1
+  utm_epsg <- ifelse(centroid_ll[2] >= 0,
+                     as.numeric(paste0("326", sprintf("%02d", utm_zone))),
[…]
