### 2026-07-14 5a99fe9
Deploying to gh-pages from @ jeremylfarrell/lakefetch@adb1951d57217de9f54761c4126708ac75baf7ff 🚀
