### 2026-07-15 832edd1
Address rOpenSci review (#762, Pakillo): v0.1.12
Follow-up to Pakillo's dead-code review comment on ropensci/software-review#762.

Removed three internal helpers that were defined and unit-tested but
never called from any production code path:
- estimate_site_depth() in R/bathymetry.R (leftover from an earlier
  attempt at per-site depth interpolation that was never wired up)
- direction_name() in R/weather_integration.R (never called)
- make_empty_osm_result() in R/lake_sources.R (added in v0.1.11 as
  scaffolding for a setTimeLimit()-based abort path that was reverted
  in the same version d
[…]
