+++ tests/testthat/test-bathymetry.R
@@ -117,41 +117,3 @@ test_that("add_lake_depth handles NA lake_osm_id", {
   expect_true(is.na(results$depth_source))
 })
 
-test_that("estimate_site_depth returns depth based on position", {
-  lake <- create_circular_lake(radius = 1000)
-
-  # Site near center
-  center_site <- sf::st_sf(
-    geometry = sf::st_sfc(sf::st_point(c(500000, 4800000)), crs = 32618)
-  )
-  depth_center <- lakefetch:::estimate_site_depth(center_site, lake, 10, 25)
-  expect_true(depth_center > 0)
-  expect_lte(depth_center, 25)
-
-  # Site near shore
-  shore_site <- sf::st_sf(
-    geometry = sf::st_sfc(sf::st_point(c(500950, 4800000)), crs = 32618)
-  )
-  depth_shore <- lakefetch:::estimate_site_depth(shore_site, lake, 10, 25)
-  expect_true(depth_shore > 0)
-  expect_true(depth_shore < depth_center)
-})
-
-test_that("estimate_site_depth handles NA depth", {
-  lake <- create_circular_lake(radius = 1000)
-  site <- sf::st_sf(
-    geometry = sf::st_sfc(sf::st_point(c(500000, 4800000)), crs = 32618)
-  )
-  result <- lakefetch:::estimate_site_depth(site, lake, NA_real_)
-  expect_true(is.na(result))
-})
-
-test_that("estimate_site_depth enforces minimum depth of 0.5m", {
-  lake <- create_circular_lake(radius = 1000)
-  # Site right at shore edge
-  edge_site <- sf::st_sf(
-    geometry = sf::st_sfc(sf::st_point(c(500999, 4800000)), crs = 32618)
-  )
-  depth <- lakefetch:::estimate_site_depth(edge_site, lake, 10, 25)
-  expect_gte(depth, 0.5)
-})
