+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: lakefetch
 Title: Calculate Fetch and Wave Exposure for Lake Sampling Points
-Version: 0.1.11
+Version: 0.1.12
 Authors@R:
     person(given = "Jeremy",
            family = "Lynch Farrell",

[…]
