### 2026-07-19 08e6d9b
Fix R-CMD-check CI not running on ropensci-software-review branch
The workflow only triggered on main/pull requests into main, but all
review commits since v0.1.4 have gone to ropensci-software-review, so
R-CMD-check had never run on any of them. This is what the
ropensci-review-bot's "Package fails continuous integration checks"
flag was catching. Add ropensci-software-review (and master, for
parity with pkgdown.yaml) to the trigger branches.

Co-Authored-By: Claude Sonnet 5 <noreply@anthropic.com>
