+++ .github/workflows/R-CMD-check.yaml
@@ -2,9 +2,9 @@ name: R-CMD-check
 
 on:
   push:
-    branches: [main]
+    branches: [main, master, ropensci-software-review]
   pull_request:
-    branches: [main]
+    branches: [main, master, ropensci-software-review]
 
 jobs:
   R-CMD-check:
