### 2026-07-19 873467b
Deploying to gh-pages from @ jeremylfarrell/lakefetch@08e6d9bbe6da40a4d0f10bd513f1f7dccb186b97 🚀
