### 2026-07-19 3131182
Add package logo
Wordmark logo displayed in the README and picked up automatically by
pkgdown for the site navbar (man/figures/logo.png).

Co-Authored-By: Claude Sonnet 5 <noreply@anthropic.com>
