### 2026-07-19 ce8f322
Deploying to gh-pages from @ jeremylfarrell/lakefetch@31311822d6f340c1f32a6ad63dba3ce2ef4495d6 🚀
