+++ llms.txt
@@ -1,5 +1,7 @@
 # lakefetch
 
+![lakefetch](reference/figures/logo.png)
+
 Calculate fetch (open water distance) and wave exposure metrics for
 freshwater lake sampling sites.
 
