### 2026-07-19 b8fbfe1
Drop plain-text README title in favor of pkgdown's auto-placed logo
pkgdown already renders man/figures/logo.png at the top of the site
automatically; the manual <img> and the "# lakefetch" heading were
both redundant with it and created visual clutter/duplication on the
homepage.

Co-Authored-By: Claude Sonnet 5 <noreply@anthropic.com>
