### 2026-07-19 145e1ed
Deploying to gh-pages from @ jeremylfarrell/lakefetch@b8fbfe17716aefa38780f357bd2fbfbd65c1f8a5 🚀
