### 2026-07-19 26e751b
Remove wordmark logo, revert to plain-text title
Holding off on a package logo for now; the wordmark was a
placeholder. Restores the default pkgdown title (no logo file means
no auto-placed corner logo either).

Co-Authored-By: Claude Sonnet 5 <noreply@anthropic.com>
