### 2026-07-19 925149b
Deploying to gh-pages from @ jeremylfarrell/lakefetch@26e751bfd7f1ae66b7455dfe9332339822c03bc5 🚀
