### 2026-07-19 2c3da58
Add concurrency groups to CI/pkgdown workflows
Neither workflow canceled superseded runs, so back-to-back pushes to
the same branch left multiple full matrix builds (including the slow
R-devel source compile) running in parallel and competing for runner
capacity. Add cancel-in-progress concurrency groups keyed on
workflow+ref so a new push cancels the previous still-running job for
that branch.

Co-Authored-By: Claude Sonnet 5 <noreply@anthropic.com>
