+++ .github/workflows/R-CMD-check.yaml
@@ -6,6 +6,10 @@ on:
   pull_request:
     branches: [main, master, ropensci-software-review]
 
+concurrency:
+  group: ${{ github.workflow }}-${{ github.ref }}
+  cancel-in-progress: true
+
 jobs:
   R-CMD-check:
     runs-on: ${{ matrix.config.os }}
