+++ .github/workflows/pkgdown.yaml
@@ -13,6 +13,10 @@ name: pkgdown
 
 permissions: read-all
 
+concurrency:
+  group: ${{ github.workflow }}-${{ github.ref }}
+  cancel-in-progress: true
+
 jobs:
   pkgdown:
     runs-on: ubuntu-latest
