### 2026-07-19 89d9622
Deploying to gh-pages from @ jeremylfarrell/lakefetch@2c3da58ff44c99b13ab9677d533c5df21826b6eb 🚀
