### 2026-07-21 2d05aaa
Post-approval housekeeping: repo transfer to ropensci org (v0.1.13)
rOpenSci review ropensci/software-review#762 was approved 2026-07-20.
This covers the reviewer bot's post-approval checklist: repoint all
GitHub links to ropensci/lakefetch after the org transfer, switch the
docs site from a self-hosted pkgdown deploy to rOpenSci's central build
at docs.ropensci.org/lakefetch, drop the local CODE_OF_CONDUCT.md in
favor of rOpenSci's default, and ignore stray local dev artifacts.

Co-Authored-By: Claude Sonnet 5 <noreply@anthropic.com>
