+++ CODE_OF_CONDUCT.md
@@ -1,31 +1,0 @@
-# Code of Conduct
-
-## Our Pledge
-
-We as contributors and maintainers pledge to make participation in this project a welcoming experience for everyone, regardless of background or identity.
-
-## Our Standards
-
-Positive behaviors include:
-- Using welcoming and inclusive language
-- Respecting differing viewpoints and experiences
-- Accepting constructive criticism gracefully
-- Focusing on what is best for the community
-
-Unacceptable behaviors include:
-- Trolling, insulting comments, or personal attacks
-- Public or private harassment
-- Publishing others' private information without permission
-- Other conduct that would be considered inappropriate in a professional setting
-
-## Scope
-
-This applies to all project spaces including the GitHub repository, issue tracker, and any other forums used for project communication.
-
-## Enforcement
-
-Instances of unacceptable behavior may be reported to the project maintainer at farrej2@rpi.edu. All reports will be reviewed and responded to appropriately.
-
-## Attribution
-
-This Code of Conduct is adapted from the [Contributor Covenant](https://www.contributor-covenant.org), version 2.1.
